Fix push for packages with multiple archive files
Golang packages usually have a vendor tarball along with the sources
archive. The existing code was picking the last archive provided by `ls`
and diffed them to figure out if the package was different.

The issue becomes worse if the files that are compared are the vendor
tarballs: the old source archives were never removed.